rename helper class to avoid name clash with other utilities
(cherry picked from commit 4333789)
